achievements

+++ SummerProject/achievements/Trait.cs
@@ -9,7 +9,7 @@ namespace SummerProject.achievements
     public class Trait
     {
         public String TraitName { get; set; }
-        public int Counter { get; set; }          
+        public float Counter { get; set; }          
 
 
         public Trait(String traitName)
